+++ src/Token.ts
@@ -1,3 +1,4 @@
+import type { Border } from "./TokenTypes/Border.js";
 import type { BoxShadow, BoxShadowList } from "./TokenTypes/BoxShadow.js";
 import type { Color } from "./TokenTypes/Color/index.js";
 import type { CubicBezier } from "./TokenTypes/CubicBezier.js";
@@ -5,6 +6,7 @@ import type { Dimension } from "./TokenTypes/Dimension.js";
 import type { Duration } from "./TokenTypes/Duration.js";
 import type { GradientList, LinearGradient, RadialGradient } from "./TokenTypes/Gradient.js";
 import type { Transition, TransitionList } from "./TokenTypes/Transition.js";
+import type { Typography } from "./TokenTypes/Typography.js";
 
 export type TokenValue =
   | string
@@ -18,6 +20,8 @@ export type TokenValue =
   | GradientList
   | Transition
   | TransitionList
+  | Typography
+  | Border
   | Dimension
   | Duration;